Make new toot visibility icon match user choice
The Toot Creator page allows users to select a visibility level for
their toots; however, this feature did not work properly. A syntax issue
involving enums was causing the logic which assigns visibility to fail,
and moreover, the icon depicting the Toot-In-Progress' visibility would
not change with the user selection. This commit fixes these issues.