Implement correct logic for Last Used Value
This MR corrects the wrong logic for determining the "Last Used" value.

Previously, the implementation relied on the last `used_at` value of the JSON array, but this approach was inconsistent and often resulted in the wrong value. This MR introduces a comparison-based logic to calculate the value accurately.

This also fixes the formatting of the dates to use the user local instead of a an hardcoded time format