Add checkbox to login page to enable moderation tools
Some servers (e.g. Pixelfed) will refuse to log in because it does not
recognize the admin scope. This might apply to other servers as well, so
make it optional when logging in. Some users who have admin privileges
may not feel comfortable giving access to Tokodon right away, so it
might function as a safety feature.