Show the loading bar even when posts exist for timeline page
We need this in some edge cases where posts are loaded, but not
completely ready yet. A good example is the account posts model, where
pinned posts are fetched after regular statuses.

A background is added to prevent the posts from being shown too early.
Setting visibility on the list view itself won't work, because it is
sometimes used to show the header (e.g. profiles) and we still need to
see that.