Extract "standalone tags" from posts, add relevant tests
This can be used to show "standalone tags" separated from the post
content. These are tags which are not embedded within the post content
themselves, and usually live on their own line at the end of a post.
This isn't given by the Mastodon API (yet?) so we have to use a bunch of
rickey regex to parse this.

I have added lots of tests for a bunch of edge cases I've seen so far,
which should help a bit.