Overhaul language selector
Now it's an OverlaySheet which is way less hacky than our custom
ComboBox reimplementation, and comes with more fringe benefits. For one,
native language names are used if available to us. Another is the
selected language is now presented in the center of the selection, so
you don't lose the selection. KItemModels is used instead of us having
another reimplementation of it.

BUG: 476285