Only subscribe to push notifications if we explicitly have the scope
This prevents Tokodon from upgrading from a non-KUnifiedPush build to
a KUnifiedPush-enabled build and spamming the user with unactionable
errors. This is caused because the push notification scope wasn't
requested before. Now that stops happening, and Tokodon now suggests
what to do when push notifications don't show up in the UI.