Add initial setup flow where required
This adds several new pages for "initial setup" in places where this is
needed. In some cases, the environment does not have a password service
so Tokodon refuses to continue in order to prevent logging in and losing
your credentials. On Android, the notifications permission is
 prompted here too.

In the ideal KDE Plasma desktop setup (or GNOME Desktop) there is no
setup.

BUG: 477927
BUG: 473787