Move multi-account home timeline name to QML, improve display
When using multiple accounts and viewing the home timeline, Tokodon puts
"Home (account name)" in the title. This is unnecessary when the drawer
is open, because it also shows which account you're on.

This removes the duplicated text when the drawer is open, and moves this
logic to QML so it's possible.