Add message to threads that some replies may be hidden
A common federation quirk is "hidden" or "missing" replies, due to
either federation being slow or the users who reply to the thread is on
servers that were blocked.

This adds a message in some situations where we can detect this
situation and suggest opening the original page to work around this
problem.