Shift around the text processing functions
This creates a new Post::processContent function which takes care of the
immediate job of taking data from JSON and putting it through the crazy
amount of processing we need to do. Removing tags and fixing the
bidirectionality of text is considered standalone enough to not put in
the Post class for brevity. The bidirectionality fixing is now
handled outside QML and happens in tandem with the rest of the
processing.