Add support for quoted posts
This is starting to become a common use case, and despite Mastodon not
implementing it (yet) Tokodon can implement it client-side.

How it works is that Tokodon looks through all the URLs of a post, and
then figures out if it's a valid post or not. To prevent hammering the
server, a basic check is performed to see if it looks like a typical
post URL. Then it's fetched from the server and displayed inline on the
timeline.