Make QtWebView optional, overhaul auth (again)
Our current authentication system uses an in-app QtWebView. However,
WebView isn't available on all systems and arguably this only makes
sense for mobile.

So now this is how the authorization should've worked before:

* On desktop systems, the external browser method is preferred. Unless
you're running on a debug build, and Tokodon will choose the auth code
method for you. This is to prevent gotchas when you forgot to install
the .desktop file and other gubbins required for that feature to work.
* On mobile systems, the in-app web browser is preferred if QtWebView
is available.
* On other systems who don't want QtWebView, they can turn it off via
USE_QTWEBVIEW and Tokodon will fall back automatically.

BUG: 483938