Fix wrong "%1 boosted" label on certain posts
The wrong author identity was selected when the boosted toot was a reply to their own post.

This is now fixed by swapping the `if`-conditions in the `identity` property to be inline with the other conditions in this file (i.e. `root.isBoosted`/`root.isReply`).

BUG: 485637

(cherry picked from commit 9e67080e6eead8bcdd6d7f8cabab65268bb324f2)