Only print out HTTP warnings for errors we don't explicitly handle
So when we make HTTP requests, by default we don't handle errors. Doing
so requires setting another callback to handle errors specifically.
We also print out a warning saying a URL 404'd and such so users can
report server errors we maybe should or shouldn't handle.

In some cases this backfires, causing errors to be printed out even if
we handle those error cases specifically. A good example is unified push
support, where we prod the server (who will return a 404 if we don't
have permissions) and this clogs up the logs. Now we consider any
explicit error callback to be considered handled and no longer print an
error in the log.

BUG: 486825