Refactor media attachment sizing to "begin" at the top
This fixes a really annoying scrolling lag when showing attachments, and
it's most likely due to layout shenanigans. To avoid this, the
attachments container now determines it's final size and the attachments
within fill the available space - instead of the other way around.

This is not backported to stable intentionally because it will most
likely break something in the meantime.

CCBUG: 490201