Fetch public servers and display them in a list for registration
For new users, it's not clear where and which servers to register if
they don't have a pre-existing Mastodon account. While Tokodon supports
registration, it still required you to know a public server URL
beforehand.

Now a nice list of public servers (courtesy of joinmastodon.org) is
displayed and you can choose from the list. You currently cannot filter
apart from the domain, but that can be added later. It's still possible
to enter a custom domain as well.