Add a new Following feed, to quickly page through your follows
Mastodon is praised for it's linear timeline, but it comes with a pretty
big caveat - users who post more often are usually the ones you see
more. Especially on servers that have a low home timeline cache limit
and users who follow a lot of people, this is troublesome. Like for me,
I only see activity for the last three days.

Fortunately Cohost has already solved this problem, so let's just copy
what they do! This adds a "Following" feed which has your followed
users on the left-hand side, and a timeline on the right. It then almost
acts like an RSS reader, where you can page through users by recent
activity to catch up on *just* their posts.

Code-wise there isn't much we need to add, except for a few
SocialGraphModel roles and a way to fetch all users.