Move sidebar to it's own file, refactor to use actions more
We eventually want to add a collapsible sidebar, but there's a few
problems to tackle first. One is that the sidebar is a hundred lines,
and adding a collapsing mode will make it larger - so we should take it
out of the Main component. Secondly is that not everything is using the
same component, nor actions so implementing the collapsing would be
harder.

Now the sidebar lives in it's own file, and everything is using actions
- even the "settings" button and others. The code is a bit simplified
too and removed some cruft we don't need anymore. The sidebar should
be functionally identical.