Disable the scrollbar interactivity for timeline pages
These are buggy because of a Qt bug we cannot fix. Using the scrollbar
with a mouse or touch just invites too much trouble, so let's disable
them ala NeoChat. It's still possible to quickly flip through pages
using PgUp/PgDown (on a keyboard), Ctrl + scroll wheel, and the
"Go to Top" button in the bottom right.

BUG: 493431