Change "Content Warning" to "Content Notice", remove warning iconography
Content warnings on Mastodon are being used (or maybe always were) as
more of a "spoiler text" or something to denote what the content is.
This isn't just done for the typical "Content Warning" (food,
eye contact, etc) but also for easy filtering - like politics, personal
life talk and more. This isn't really deserving of the warning
iconography we give it, and more people are against that in upstream
Mastodon as well.