Add support for displaying authorship in preview cards
This overhauls the look of preview cards, and this allows us to display
authorship information. The biggest change is now the card is split into
two separate controls, one for the actual card preview and the other for
authorship. If Fediverse authorship is provided, an account link is used
instead of the website link to the author.

These buttons now have visible hover states, and are still keyboard
accessible.