Support new notification events in v4.3.0, along with report types
Whilst the notification delegates are limited, they are now shown in the
notification pane. This includes the "report" notification for admins
(albiet you still need to open up the tools manually to see it for now)
and the two new events in v4.3.0: severance events and moderation
warnings. The details shown can be expanded in the future.

BUG: 485168