Simplify read markers
Instead of trying to load the timeline starting from the read marker,
keep the old behavior of loading the timeline from the last message and
instead show the read marker inside the timeline.

This prevents issues like not being able to load any message when the
marker points to an invalid read marker.