Fix that one annoying visual bug where the reply/boost user doesn't show
This is the "purple icon" bug where sometimes, the identity of whoever
boosted or replied to a post is missing. This usually happens during the
initial timeline load, because we could be fetching the account info
from the server. If this happens, we need to make sure to update the
relevant roles on the timeline.

(cherry picked from commit 981b88b3c253276cb6b44ef77ab6d5f2574cdc3f)