Fix that one annoying visual bug where the reply/boost user doesn't show
This is the "purple icon" bug where sometimes, the identity of whoever
boosted or replied to a post is missing. This usually happens during the
initial timeline load, because we could be fetching the account info
from the server. If this happens, we need to make sure to update the
relevant roles on the timeline.