Fix files not saving on Android
It turns out to be two little nagging bugs. First, we are checking if
the temporary file is readable on Android (it never is) and fails.
Secondly, we need to not convert the content:// URI to a local file URI
since that fails. Finally, we need to enable direct write fallback for
QSaveFile because yet again, we cannot write anywhere else in the
directory on Android.

BUG: 496010