Improve quote post detection
The first issue I fixed is that not all URLs were considered in the post
body, due to an oversight when I originally added this.

Secondly, quoted posts should now work when the original poster is also
mentioned in the post. This was due to all URLs being rewritten with the
internal account:/ URI, even if they weren't a mention.

(cherry picked from commit cdc97e314114969ad636484b7ab0a8c43b046199)