Don't load the media player until it's actually needed
This *significantly* improves the performance of scrolling through the
timeline. Media players from MpvQt and QtMultimedia are extremely heavy
and take a while to initialize.

While this matters the most to people who prefer GIFs to not auto-play,
this still has the beneficial effect for long-form videos which never
auto-play.

This has no visible behavioral change, however I did fix a bug where
GIFs did not loop when built with the QtMultimedia backend.