Improve the look and feel of image loading, and hiding media
First of all, a BusyIndicator is added to indicate that an image is
indeed loading. But the more obvious visual change is now it will
smoothly animate from the loading state (which is a nice blurhash) to
it's actual form. This also works when hiding/showing media, but unlike
certain other clients - the sensitive media isn't shown for a split
second.

(cherry picked from commit 088dc03c3699c4cc19117e5d78676c9da8876fd6)