Add adaptive launcher logo for android
On android the launcher logo is not looking native. Added adaptive logo for the launcher. In the logo the shadows are removed to fulfill the platform design.
This changes were not tested, just checked the .apk from the ci that the icons are in.

(cherry picked from commit 2960d72811a4a5be5c2e3fb2e6cdf9b31f9cf8d9)