Add error log page in the settings
This gives you the error string and the URL that failed in a visible
place in the UI. The last five error messages are saved across restarts,
making this extremely useful for bug reports.

Currently this is only really used for network errors but could be
expanded for other types of errors later.

CCBUG: 485807