Completely redo how errors are shown
Previously, Tokodon would report unhandled errors. But in actuality,
most API calls are harmless if they error out (for example, boosting a
post, setting a read marker and so on.) It should instead be the other
way around, call-sites should decide themselves whether to elevate the
error.

In all cases, the error is *always* logged so it can be fixed in Tokodon
or in server software.