Add loader back for AccountHeader
This semantically reverts 650c489da20b6be87186cdeed1d29c2c90284f0c by
adding a Loader back. But instead for both pages, it only adds a Loader
for the initial page in the StackLayout - the one used for Posts and
Replies.

Also some changes were made to check for a null identity, such as in the
title for this page.