Allow pinning favorite lists to the sidebar for easy access
For commonly used lists, it takes two taps in order to actually access
it which is really annoying. Now you can pin them as favorites, which
will show up in the sidebar like any other feed.

On the technical side, it's ensured that these lists are valid and any
deleted ones will be removed. The favorites are stored in the usual
account config.