Fix crash when receiving unknown notifications
There's a Q_UNREACHABLE() in the default case, but this turns out to be
pretty bad. The server can send unknown notifications behind our back or
we don't send it post it as a notification yet.

This fixes a regular crash I've seen when receiving push notifications.
I also added a warning I can grep for in the logs whenever a
notification type is unknown.

BUG: 497915