Always show the reply count for posts
Up until recently you were able to tell how many replies a post has by
if the post icon was filled or not. Then we killed our custom post icon,
which also got rid of this indicator.

Now the reply count is always shown regardless if you want to see post
stats or not. This is also how Mastodon Web handles this.