Remove MpvQt support
We have had optional QtMultimedia support for a while now, and I feel
like MpvQt isn't as needed as it used to be. Let's remove it and
simplify our multimedia dependencies a bit.

Note that this doesn't affect future wishlist items like PeerTube
support. That can still be done under QtMultimedia.