Add a prompt to remind you to add alt text to media
This is a common feature in several Mastodon clients and even BlueSky,
which is to prompt the user before posting media without alt text. I
always make a habit of this, so this is even useful for me so I don't
forget.

Of course, if you really intend to not use alt text you can skip the
reminder. This dialog can be turned off if you find it annoying.