Improve refreshing the main timeline, other misc fixes
Currently refreshing the main timeline involves throwing everything away
and starting fresh, but we don't need to do that. Now we use the
pagination to load all newer posts, and we're a bit smarter about it
too.

An additional bug that I think I fixed is receiving duplicate posts. Now
fetchedTimeline and handleEvent shouldn't add posts we have already.

A bug that I ran into but not sure if it's ever resulted in a crash is
not checking if posts is empty before calling first() on it, but good to
fix anyway.

CCBUG: 495996
BUG: 477420