Don't reload the current page when clicking on it in the navigation
Previously if you clicked on a page in the navigation - e.g "Home" -
then Tokodon would load the *whole* page again. Not only is this super
wasteful, but it's also a performance drain and bad UX.

Instead, don't allow reloading to the *same* page you were already on.
And start making a repeated tap do something better. Right now
for example tapping on Home again will just return you to the top.

CCBUG: 497697