Allow reading out post content with a screen reader, improve output
Now it's actually possible to read out the post's textual content with
a screen reader, but it's not as easy as putting it in the tab order. We
have to also strip out all of the invisible, non-human readable HTML or
else the screen reader still spend *minutes* speaking about an anchor
tag or something stupid like that.