Allow opening links via tokodon:// url scheme
This URL scheme is registered and used for oauth codes, but we can
expand this to be used alongside the existing handling of web+ap:// and
https://.

One of the ways this can be useful is for opening specifically *Tokodon*
instead of maybe another Mastodon client that also handles web+ap like
Tuba. For example, using tokodon:// in the StreetPass extension.