Hide Purpose's own "Copy to Clipboard" function
This was added recently to Purpose, and it's not too useful for us. When
you try to paste it, it downloads the page from your Mastodon server
which doesn't work as Mastodon depends on JavaScript. And having two
ways to copy stuff to the clipboard is just confusing, so we should hide
this plugin.