Fix quote posts always showing a poll
First of all, this fixes a bug where polls would show up in the quoted
post despite it not working correctly. You usually end up tapping into
the post itself when selecting stuff, and voting doesn't work.

Second, this doesn't show the full poll - because that's unnecessary -
and instead shows a minimized version of the poll.