Stop registering for push notifications despite them being disabled
We previously depended on a non-success HTTP status for detecting if we
had a previous subscription or not. That's not enough, apparently! The
server could also respond with an OK, and return a JSON object
containing an error too, so we have to handle that.

This should fix the odd situation where you have KUnifiedPush, disabled
notifications for the account and Tokodon registering you anyway.
Additionally, we would also *unregister* you the next time you
booted up Tokodon, fun!