Change some notification delegates to a smaller post preview
This one has a couple of differences to the main post delegate. Notably,
the post info bar is gone - because all it does is duplicate your own
information too much. The image attachments are force cropped to reduce
their height even more.

These replace the favorite and boost delegates, which are only your own
posts. But this delegate could also be used for the status composer
preview.