Add better tooltips for disabled poll and attachment buttons
This prevents confusion to people new to Mastodon, and may not be aware
that they don't allow showing both media and polls at the same time. We
previously disabled the button (which matches upstream Mastodon Web) but
don't actually give an explanation in-application as to why.

Now the tooltips for these buttons match their disabled state.

CCBUG: 504946