Add better error messages for some edge cases during authentication
For example, I still had an account for my old server on this machine,
and it showed a blank error message in the UI. The reason why it was
blank is that Nginx served me an HTML error message that *technically*
isn't error.

That still isn't the greatest, it should show me "bad gateway" but it's
something better I guess.