Prevent a null access in certain timelines
I don't know why this happens exactly, but I have a hunch. It's possible
for us to end up with null pointers because of the post->hidden() check
and that's not really checked in the erasure. I added an early enough
check in there to hopefully stop that from happening.

BUG: 504215


(cherry picked from commit da024d8a55193d996d514dd0691808a3a71ad157)

Co-authored-by: Joshua Goins <josh@redstrate.com>