Fix "View Website" button sometimes trying to open in the qrc scheme
This is because with some configs, the InstanceUri config is stored
without a scheme. Because Qt is wonderful, it assumes that URLs without
a scheme is actually supposed to be QRC files.

To stop this, I made instanceUri return a QUrl that automatically
corrects itself to https if an explicit scheme isn't set.

This should fix various links throughout the application - such as the
link on the login issue page.