Improve reliability of canFetchMore in our models
None of them checked if we are currently loading, which could be a
problem depending on if their fetch function early returns. (But it's
not a bad idea to check that early for the ones that do.)

Also, some models need to check if they have any next links to follow.